3DS auto pay issue fixed
