+++ app/Console/Commands/RenewalCron.php
@@ -2,7 +2,8 @@
 
 namespace App\Console\Commands;
 
-use App\Http\Controllers\Common\CronController;
+use App\Http\Controllers\Subscription\SubscriptionController;
+use App\Http\Controllers\ConcretePostSubscriptionHandleController; // Import the concrete controller
 use Illuminate\Console\Command;
 
 class RenewalCron extends Command
@@ -28,8 +29,16 @@ class RenewalCron extends Command
      */
     public function handle()
     {
-        $controller = new CronController();
+        // Create an instance of ConcretePostSubscriptionHandleController
+        $concreteController = app()->make(ConcretePostSubscriptionHandleController::class);
+
+        // Pass the concrete controller instance to CronController constructor
+        $controller = new SubscriptionController($concreteController);
+
+        // Call the method to perform auto-renewal
         $controller->autoRenewal();
+
+        // Output success message
         $this->info('renewal:cron Command Run successfully!');
     }
 }
